app/services/audio_plagiarism.py
"""
Audio plagiarism detection service using Music Information Retrieval (MIR) techniques
"""
import numpy as np
import librosa
from scipy.spatial.distance import cosine
from typing import Optional, BinaryIO
import logging

logger = logging.getLogger(__name__)


def score_plagiarism(file_like: BinaryIO, filename: str = "audio") -> Optional[float]:
    """
    Analyze audio for potential plagiarism using MFCC feature comparison

    Args:
        file_like: Binary file-like object containing audio data
        filename: Name of the file for logging purposes

    Returns:
        Float between 0.0 and 1.0 indicating plagiarism likelihood
        (0.0 = no similarity, 1.0 = identical)
    """
    try:
        # Load audio file
        y, sr = librosa.load(file_like, sr=None, mono=True,
                             duration=30)  # Analyze first 30 seconds

        if len(y) == 0:
            logger.warning(f"Empty audio file: {filename}")
            return 0.0

        # Extract MFCC features (Mel-frequency cepstral coefficients)
        mfcc = librosa.feature.mfcc(y=y, sr=sr, n_mfcc=20)

        # Calculate mean across time axis for each coefficient
        mfcc_mean = mfcc.mean(axis=1)

        # Extract additional features for better comparison
        chroma = librosa.feature.chroma_stft(y=y, sr=sr).mean(axis=1)
        spectral_centroid = librosa.feature.spectral_centroid(
            y=y, sr=sr).mean()
        tempo, _ = librosa.beat.beat_track(y=y, sr=sr)

        # Combine features into a single vector
        feature_vector = np.concatenate([
            mfcc_mean,
            chroma,
            [spectral_centroid, tempo]
        ])

        # Normalize features
        feature_vector = feature_vector / \
            (np.linalg.norm(feature_vector) + 1e-8)

        # TODO: In a real implementation, compare against a database of known tracks
        # For now, we'll simulate plagiarism detection with a simple heuristic

        # Calculate a pseudo-similarity score based on feature characteristics
        # This is a placeholder - in production, you'd compare against a reference database
        similarity_score = _calculate_similarity_heuristic(
            feature_vector, mfcc, chroma)

        logger.info(
            f"Plagiarism analysis completed for {filename}: score={similarity_score:.3f}")

        return float(np.clip(similarity_score, 0.0, 1.0))

    except Exception as e:
        logger.error(
            f"Error analyzing audio plagiarism for {filename}: {str(e)}")
        return 0.0


def _calculate_similarity_heuristic(feature_vector: np.ndarray, mfcc: np.ndarray, chroma: np.ndarray) -> float:
    """
    Calculate a heuristic similarity score based on audio features

    This is a placeholder implementation. In a real system, you would:
    1. Compare against a database of known copyrighted works
    2. Use more sophisticated similarity metrics
    3. Implement fingerprinting techniques like Shazam's algorithm
    """

    # Analyze feature variance (low variance might indicate repetitive/simple content)
    mfcc_variance = np.var(mfcc, axis=1).mean()
    chroma_variance = np.var(chroma, axis=1).mean()

    # Analyze spectral characteristics
    spectral_complexity = np.std(feature_vector[:20])  # MFCC complexity
    harmonic_complexity = np.std(feature_vector[20:32])  # Chroma complexity

    # Simple heuristic: higher complexity = lower plagiarism likelihood
    # This is just for demonstration - real implementation would be much more sophisticated
    complexity_score = (
        spectral_complexity + harmonic_complexity + mfcc_variance + chroma_variance) / 4

    # Invert and normalize to get plagiarism likelihood
    # Higher complexity = lower plagiarism score
    plagiarism_score = max(0.0, 0.8 - complexity_score * 2.0)

    # Add some randomness to simulate real-world variability
    # In production, this would be replaced with actual database comparisons
    import random
    random.seed(int(np.sum(feature_vector) * 1000) %
                1000)  # Deterministic but varies by content
    noise = random.uniform(-0.1, 0.1)

    return plagiarism_score + noise


def extract_audio_fingerprint(file_like: BinaryIO) -> Optional[np.ndarray]:
    """
    Extract a compact audio fingerprint for efficient comparison

    This could be used for building a reference database of copyrighted works
    """
    try:
        y, sr = librosa.load(file_like, sr=22050, mono=True, duration=30)

        # Extract key features for fingerprinting
        mfcc = librosa.feature.mfcc(y=y, sr=sr, n_mfcc=12)
        chroma = librosa.feature.chroma_stft(y=y, sr=sr)

        # Create compact fingerprint
        fingerprint = np.concatenate([
            mfcc.mean(axis=1),
            mfcc.std(axis=1),
            chroma.mean(axis=1)
        ])

        return fingerprint

    except Exception as e:
        logger.error(f"Error extracting audio fingerprint: {str(e)}")
        return None


def compare_fingerprints(fp1: np.ndarray, fp2: np.ndarray) -> float:
    """
    Compare two audio fingerprints and return similarity score
    """
    try:
        # Normalize fingerprints
        fp1_norm = fp1 / (np.linalg.norm(fp1) + 1e-8)
        fp2_norm = fp2 / (np.linalg.norm(fp2) + 1e-8)

        # Calculate cosine similarity
        similarity = 1.0 - cosine(fp1_norm, fp2_norm)

        return max(0.0, similarity)

    except Exception as e:
        logger.error(f"Error comparing fingerprints: {str(e)}")
        return 0.0
